all endpoints done; structure set ...
