--- mock-server/src/services/shared/shipment-rates.ts
@@ -1,8 +1,0 @@
-import {Request, Response} from 'express'
-
-/**
- * Get rates for a shipment.
- */
-export const shipmentRates = (_: Request, res: Response): void => {
-  res.json({});
-}
