--- mock-server/src/utils/range.ts
@@ -1,1 +1,0 @@
-export const range = (n: number): number[] => [...Array(n)].map((_, i) => i + 1)
