+++ mock-server/src/utils/types.ts
